requirements.txt
aiogram>=3.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
matplotlib==3.8.2
pandas==2.1.4 